+++ src/main.ts
@@ -35,7 +35,6 @@ const boostsrap = async () => {
         const model = ModelFactory.create(AppModel);
         model.setup();
         const scene = new ModelScene(model);
-        console.log('okk');
         Manager.changeScene(scene);
     });
 }
