+++ src/ui/scenes/loader.scene.ts
@@ -7,8 +7,6 @@ export class LoaderScene extends PixiContainer implements SceneInterface {
     private _loadingBar: LoadingBarContainer;
     constructor() {
         super();
-        //this.position.x = Manager.width/2;
-        //this.position.y = Manager.height/2;
         const parentWidth = Manager.width;
         const parentHeight = Manager.height;
 
@@ -21,7 +19,7 @@ export class LoaderScene extends PixiContainer implements SceneInterface {
         this._loadingBar.scaleProgress(progress);
     }
 
-    update(framesPassed: number): void { framesPassed = 0; }
+    update(_framesPassed: number): void { }
     resize(screenWidth: number, screenHeight: number): void { 
         this.width = screenWidth; 
         this.height = screenHeight;
