--- src/ui/sprites/pixi-logo.sprite.ts
@@ -1,8 +1,0 @@
-import { PixiTexture, PixiSprite } from "../../plugins/engine";
-
-export class PixiLogoSprite extends PixiSprite {
-    constructor(source: string) {
-        const pixiTexture = PixiTexture.from(source);
-        super(pixiTexture);
-    }
-}
