--- src/ui/textures/pixi-logo.texture.ts
@@ -1,8 +1,0 @@
-import { PixiTexture, PixiBaseTexture } from "../../plugins/engine";
-
-export class PixiLogoTexture extends PixiTexture {
-    constructor(source: string) {
-        const baseTexture = PixiBaseTexture.from(source);
-        super(baseTexture);
-    }
-}
